+++ src/main/java/com/keymouseshare/bean/ScreenCoordinate.java
@@ -3,14 +3,14 @@ package com.keymouseshare.bean;
 public class ScreenCoordinate {
     public String ip;
     public String screenName;
-    public double x;
-    public double y;
+    public int x;
+    public int y;
 
 
     public ScreenCoordinate() {
     }
 
-    public ScreenCoordinate(String ip, String screenName, double x, double y) {
+    public ScreenCoordinate(String ip, String screenName, int x, int y) {
         this.ip = ip;
         this.screenName = screenName;
         this.x = x;
