+++ src/main/java/com/keymouseshare/bean/ScreenInfo.java
@@ -8,19 +8,19 @@ package com.keymouseshare.bean;
 public class ScreenInfo {
     private String deviceIp;
     private String screenName;
-    private double width;
-    private double height;
-    private double dx; // 本地屏幕坐标
-    private double dy; // 本地屏幕坐标
-    private double vx; // 虚拟桌面屏幕坐标
-    private double vy; // 虚拟桌面屏幕坐标
-    private double mx; // 屏幕配置坐标
-    private double my; // 屏幕配置坐标
+    private int width;
+    private int height;
+    private int dx; // 本地屏幕坐标
+    private int dy; // 本地屏幕坐标
+    private int vx; // 虚拟桌面屏幕坐标
+    private int vy; // 虚拟桌面屏幕坐标
+    private int mx; // 屏幕配置坐标
+    private int my; // 屏幕配置坐标
 
     public ScreenInfo() {}
 
 
-    public ScreenInfo(String deviceIp, String screenName, double width, double height, double dx, double dy) {
+    public ScreenInfo(String deviceIp, String screenName, int width, int height, int dx, int dy) {
         this.deviceIp = deviceIp;
         this.screenName = screenName;
         this.width = width;
@@ -30,7 +30,7 @@ public class ScreenInfo {
     }
 
 
-    public ScreenInfo(String deviceIp, String screenName, double width, double height) {
+    public ScreenInfo(String deviceIp, String screenName, int width, int height) {
         this.deviceIp = deviceIp;
         this.screenName = screenName;
         this.width = width;
@@ -42,51 +42,51 @@ public class ScreenInfo {
     public String getScreenName() { return screenName; }
     public void setScreenName(String screenName) { this.screenName = screenName; }
 
-    public double getWidth() { return width; }
+    public int getWidth() { return width; }
     public void setWidth(int width) { this.width = width; }
 
-    public double getHeight() { return height; }
+    public int getHeight() { return height; }
     public void setHeight(int height) { this.height = height; }
 
     public String getDeviceIp() { return deviceIp; }
     public void setDeviceIp(String deviceIp) { this.deviceIp = deviceIp; }
 
-    public double getDx() { return dx; }
+    public int getDx() { return dx; }
 
-    public void setDx(double dx) { this.dx = dx; }
+    public void setDx(int dx) { this.dx = dx; }
 
-    public double getDy() { return dy; }
+    public int getDy() { return dy; }
 
-    public void setDy(double dy) { this.dy = dy; }
+    public void setDy(int dy) { this.dy = dy; }
 
-    public double getVx() { return vx; }
+    public int getVx() { return vx; }
 
-    public void setVx(double vx) { this.vx = vx; }
+    public void setVx(int vx) { this.vx = vx; }
 
-    public double getVy() { return vy; }
+    public int getVy() { return vy; }
 
-    public void setVy(double vy) { this.vy = vy; }
+    public void setVy(int vy) { this.vy = vy; }
 
-    public void setMx(double mx) {
+    public void setMx(int mx) {
         this.mx = mx;
     }
 
-    public double getMx() {
+    public int getMx() {
         return mx;
     }
 
-    public void setMy(double my) {
+    public void setMy(int my) {
         this.my = my;
     }
 
-    public double getMy() {
+    public int getMy() {
         return my;
     }
 
-    public boolean localContains(double globalX, double globalY) {
+    public boolean localContains(int globalX, int globalY) {
         return globalX >= dx && globalX < dx + width && globalY >= dy && globalY < dy + height;
     }
-    public boolean virtualContains(double globalX, double globalY) {
+    public boolean virtualContains(int globalX, int globalY) {
         return globalX >= vx && globalX < vx + width && globalY >= vy && globalY < vy + height;
     }
 }
